PB-675 move certificate type
